fix(coap): invert the condition for when to build OSCRE context
This ensures that code path to ariel-os-random is only traversed when
something that selects the random module is actually active.

Not setting the otherwise unused variables avoids both linter cases and
ensures clarity to the reader.

It is a difference in behavior exclusively in the case of running CoAP
with no server support (and also configured server security setup), but
for that case, the handler that is exposed is the trivial "everything is
4.04" already.

The change to not build a .well-known/core resource on client-only
system ensures that any future more elaborate .well-known/core
implementation from the coap-handler-implementations can not be accessed
by clients (as it would have been enforced by policy earlier), and is
generally an improvement for client-only sytems because it sheds code.

See-also: https://github.com/ariel-os/ariel-os/pull/1618#discussion_r2626275091
Signed-off-by: chrysn <chrysn@fsfe.org>